+++ Userland/Libraries/LibJS/Runtime/Intl/NumberFormatConstructor.h
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021, Tim Flynn <trflynn89@serenityos.org>
+ * Copyright (c) 2021-2022, Tim Flynn <trflynn89@serenityos.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
 #pragma once
 
+#include <LibJS/Runtime/Intl/NumberFormat.h>
 #include <LibJS/Runtime/NativeFunction.h>
 
 namespace JS::Intl {
@@ -26,5 +27,9 @@ private:
 
     JS_DECLARE_NATIVE_FUNCTION(supported_locales_of);
 };
+
+ThrowCompletionOr<NumberFormat*> initialize_number_format(GlobalObject& global_object, NumberFormat& number_format, Value locales_value, Value options_value);
+ThrowCompletionOr<void> set_number_format_digit_options(GlobalObject& global_object, NumberFormatBase& intl_object, Object const& options, int default_min_fraction_digits, int default_max_fraction_digits, NumberFormat::Notation notation);
+ThrowCompletionOr<void> set_number_format_unit_options(GlobalObject& global_object, NumberFormat& intl_object, Object const& options);
 
 }
